+++ src/net_main.c
@@ -46,7 +46,7 @@ static qboolean	listening = false;
 qboolean	slistInProgress = false;
 qboolean	slistSilent = false;
 qboolean	slistLocal = true;
-static double	slistStartTime;
+static float	slistStartTime;
 static int		slistLastShown;
 
 static void Slist_Send(void);
@@ -90,7 +90,7 @@ int	net_driverlevel;
 
 static PollProcedure *pollProcedureList = NULL;
 
-double			net_time;
+float			net_time;
 int hostCacheCount = 0;
 hostcache_t hostcache[HOSTCACHESIZE];
 
@@ -100,7 +100,7 @@ SetNetTime
 
 ===================
 */
-double SetNetTime(void)
+float SetNetTime(void)
 {
 	net_time = Sys_FloatTime();
 	return net_time;
@@ -455,7 +455,7 @@ NET_CheckNewConnections
 
 struct
 {
-	double	time;
+	float	time;
 	int		op;
 	long	session;
 } vcrConnect;
@@ -534,7 +534,7 @@ returns -1 if connection is invalid
 
 struct
 {
-	double	time;
+	float	time;
 	int		op;
 	long	session;
 	int		ret;
@@ -622,7 +622,7 @@ returns -1 if the connection died
 */
 struct
 {
-	double	time;
+	float	time;
 	int		op;
 	long	session;
 	int		r;
@@ -727,7 +727,7 @@ qboolean NET_CanSendMessage (qsocket_t *sock)
 
 int NET_SendToAll(sizebuf_t *data, int blocktime)
 {
-	double		start;
+	float		start;
 	int			i;
 	int			count = 0;
 	qboolean	state1 [MAX_SCOREBOARD];
@@ -957,7 +957,7 @@ void NET_Poll(void)
 	}
 }
 
-void SchedulePollProcedure(PollProcedure *proc, double timeOffset)
+void SchedulePollProcedure(PollProcedure *proc, float timeOffset)
 {
 	PollProcedure *pp, *prev;
